more serial work, remove old driver files
