+++ android/package/src/com/hoho/android/usbserial/driver/Ch34xSerialDriver.java
