--- android/package/src/com/hoho/android/usbserial/driver/CommonUsbSerialDriver.java
@@ -1,172 +1,0 @@
-/* Copyright 2013 Google Inc.
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301,
- * USA.
- *
- * Project home page: http://code.google.com/p/usb-serial-for-android/
- */
-
-package com.hoho.android.usbserial.driver;
-
-import android.hardware.usb.UsbDevice;
-import android.hardware.usb.UsbDeviceConnection;
-
-import java.io.IOException;
-
-/**
- * A base class shared by several driver implementations.
- *
- * @author mike wakerly (opensource@hoho.com)
- */
-abstract class CommonUsbSerialDriver implements UsbSerialDriver {
-
-    public static final int DEFAULT_READ_BUFFER_SIZE = 16 * 1024;
-    public static final int DEFAULT_WRITE_BUFFER_SIZE = 16 * 1024;
-
-    protected final UsbDevice   mDevice;
-    protected final Object      mReadBufferLock = new Object();
-    protected final Object      mWriteBufferLock = new Object();
-
-    protected UsbDeviceConnection mConnection = null;
-
-    private int _permissionStatus = permissionStatusRequestRequired;
-
-    /** Internal read buffer.  Guarded by {@link #mReadBufferLock}. */
-    protected byte[] mReadBuffer;
-
-    /** Internal write buffer.  Guarded by {@link #mWriteBufferLock}. */
-    protected byte[] mWriteBuffer;
-
-    public CommonUsbSerialDriver(UsbDevice device) {
-        mDevice = device;
-
-        mReadBuffer = new byte[DEFAULT_READ_BUFFER_SIZE];
-        mWriteBuffer = new byte[DEFAULT_WRITE_BUFFER_SIZE];
-    }
-
-    @Override
-    public void setConnection(UsbDeviceConnection connection) {
-        mConnection = connection;
-    }
-
-    @Override
-    public int permissionStatus() {
-        return _permissionStatus;
-    }
-
-    @Override
-    public void setPermissionStatus(int permissionStatus) {
-        _permissionStatus = permissionStatus;
-    }
-
-    /**
-     * Returns the currently-bound USB device.
-     *
-     * @return the device
-     */
-    @Override
-    public final UsbDevice getDevice() {
-        return mDevice;
-    }
-
-    /**
-     * Returns the currently-bound USB device connection.
-     *
-     * @return the device connection
-     */
-    @Override
-    public final UsbDeviceConnection getDeviceConnection() {
-        return mConnection;
-    }
-
-
-
-    /**
-     * Sets the size of the internal buffer used to exchange data with the USB
-     * stack for read operations.  Most users should not need to change this.
-     *
-     * @param bufferSize the size in bytes
-     */
-    public final void setReadBufferSize(int bufferSize) {
-        synchronized (mReadBufferLock) {
-            if (bufferSize == mReadBuffer.length) {
-                return;
-            }
-            mReadBuffer = new byte[bufferSize];
-        }
-    }
-
-    /**
-     * Sets the size of the internal buffer used to exchange data with the USB
-     * stack for write operations.  Most users should not need to change this.
-     *
-     * @param bufferSize the size in bytes
-     */
-    public final void setWriteBufferSize(int bufferSize) {
-        synchronized (mWriteBufferLock) {
-            if (bufferSize == mWriteBuffer.length) {
-                return;
-            }
-            mWriteBuffer = new byte[bufferSize];
-        }
-    }
-
-    @Override
-    public abstract void open() throws IOException;
-
-    @Override
-    public abstract void close() throws IOException;
-
-    @Override
-    public abstract int read(final byte[] dest, final int timeoutMillis) throws IOException;
-
-    @Override
-    public abstract int write(final byte[] src, final int timeoutMillis) throws IOException;
-
-    @Override
-    public abstract void setParameters(
-            int baudRate, int dataBits, int stopBits, int parity) throws IOException;
-
-    @Override
-    public abstract boolean getCD() throws IOException;
-
-    @Override
-    public abstract boolean getCTS() throws IOException;
-
-    @Override
-    public abstract boolean getDSR() throws IOException;
-
-    @Override
-    public abstract boolean getDTR() throws IOException;
-
-    @Override
-    public abstract void setDTR(boolean value) throws IOException;
-
-    @Override
-    public abstract boolean getRI() throws IOException;
-
-    @Override
-    public abstract boolean getRTS() throws IOException;
-
-    @Override
-    public abstract void setRTS(boolean value) throws IOException;
-
-    @Override
-    public boolean purgeHwBuffers(boolean flushReadBuffers, boolean flushWriteBuffers) throws IOException {
-        return !flushReadBuffers && !flushWriteBuffers;
-    }
-
-}
-
