+++ android/package/src/com/hoho/android/usbserial/driver/CommonUsbSerialPort.java
