+++ android/package/src/com/hoho/android/usbserial/driver/Cp21xxSerialDriver.java
