+++ android/package/src/com/hoho/android/usbserial/driver/GsmModemSerialDriver.java
