+++ android/package/src/com/hoho/android/usbserial/driver/SerialTimeoutException.java
