+++ android/package/src/com/hoho/android/usbserial/driver/UsbId.java
@@ -1,22 +1,9 @@
-/* Copyright 2012 Google Inc.
+/* Copyright 2011-2013 Google Inc.
+ * Copyright 2013 mike wakerly <opensource@hoho.com>
  *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301,
- * USA.
- *
- * Project home page: http://code.google.com/p/usb-serial-for-android/
+ * Project home page: https://github.com/mik3y/usb-serial-for-android
  */
+
 package com.hoho.android.usbserial.driver;
 
 /**
@@ -31,58 +18,39 @@ public final class UsbId {
 
     public static final int VENDOR_FTDI = 0x0403;
     public static final int FTDI_FT232R = 0x6001;
-    public static final int FTDI_FT231X = 0x6015;
+    public static final int FTDI_FT2232H = 0x6010;
+    public static final int FTDI_FT4232H = 0x6011;
+    public static final int FTDI_FT232H = 0x6014;
+    public static final int FTDI_FT231X = 0x6015; // same ID for FT230X, FT231X, FT234XD
 
-    public static final int VENDOR_PX4    = 0x26AC;
-    public static final int DEVICE_PX4FMU = 0x11;
-
-    public static final int VENDOR_ATMEL = 0x03EB;
-    public static final int ATMEL_LUFA_CDC_DEMO_APP = 0x2044;
-
-    public static final int VENDOR_ARDUINO = 0x2341;
-    public static final int ARDUINO_UNO = 0x0001;
-    public static final int ARDUINO_MEGA_2560 = 0x0010;
-    public static final int ARDUINO_SERIAL_ADAPTER = 0x003b;
-    public static final int ARDUINO_MEGA_ADK = 0x003f;
-    public static final int ARDUINO_MEGA_2560_R3 = 0x0042;
-    public static final int ARDUINO_UNO_R3 = 0x0043;
-    public static final int ARDUINO_MEGA_ADK_R3 = 0x0044;
-    public static final int ARDUINO_SERIAL_ADAPTER_R3 = 0x0044;
-    public static final int ARDUINO_LEONARDO = 0x8036;
-
-    public static final int VENDOR_VAN_OOIJEN_TECH = 0x16c0;
-    public static final int VAN_OOIJEN_TECH_TEENSYDUINO_SERIAL = 0x0483;
-
-    public static final int VENDOR_LEAFLABS = 0x1eaf;
-    public static final int LEAFLABS_MAPLE = 0x0004;
-    
-    public static final int VENDOR_SILAB = 0x10c4;
-    public static final int SILAB_CP2102 = 0xea60;
+    public static final int VENDOR_SILABS = 0x10c4;
+    public static final int SILABS_CP2102 = 0xea60; // same ID for CP2101, CP2103, CP2104, CP2109
+    public static final int SILABS_CP2105 = 0xea70;
+    public static final int SILABS_CP2108 = 0xea71;
 
     public static final int VENDOR_PROLIFIC = 0x067b;
-    public static final int PROLIFIC_PL2303 = 0x2303;
+    public static final int PROLIFIC_PL2303 = 0x2303;   // device type 01, T, HX
+    public static final int PROLIFIC_PL2303GC = 0x23a3; // device type HXN
+    public static final int PROLIFIC_PL2303GB = 0x23b3; // "
+    public static final int PROLIFIC_PL2303GT = 0x23c3; // "
+    public static final int PROLIFIC_PL2303GL = 0x23d3; // "
+    public static final int PROLIFIC_PL2303GE = 0x23e3; // "
+    public static final int PROLIFIC_PL2303GS = 0x23f3; // "
 
-    public static final int VENDOR_UBLOX = 0x1546;
-    public static final int DEVICE_UBLOX_5 = 0x01a5;
-    public static final int DEVICE_UBLOX_6 = 0x01a6;
-    public static final int DEVICE_UBLOX_7 = 0x01a7;
-    public static final int DEVICE_UBLOX_8 = 0x01a8;
+    public static final int VENDOR_GOOGLE = 0x18d1;
+    public static final int GOOGLE_CR50 = 0x5014;
 
-    public static final int VENDOR_OPENPILOT = 0x20A0;
-    public static final int DEVICE_REVOLUTION = 0x415E;
-    public static final int DEVICE_OPLINK = 0x415C;
-    public static final int DEVICE_SPARKY2 = 0x41D0;
-    public static final int DEVICE_CC3D = 0x415D;
+    public static final int VENDOR_QINHENG = 0x1a86;
+    public static final int QINHENG_CH340 = 0x7523;
+    public static final int QINHENG_CH341A = 0x5523;
 
-    public static final int VENDOR_ARDUPILOT_CHIBIOS1 = 0x0483;
-    public static final int VENDOR_ARDUPILOT_CHIBIOS2 = 0x1209;
-    public static final int DEVICE_ARDUPILOT_CHIBIOS =  0x5740;
+    public static final int VENDOR_UNISOC = 0x1782;
+    public static final int FIBOCOM_L610 = 0x4D10;
+    public static final int FIBOCOM_L612 = 0x4D12;
 
-    public static final int VENDOR_DRAGONLINK = 0x1FC9;
-    public static final int DEVICE_DRAGONLINK = 0x0083;
 
     private UsbId() {
-        throw new IllegalAccessError("Non-instantiable class.");
+        throw new IllegalAccessError("Non-instantiable class");
     }
 
 }
