+++ android/package/src/com/hoho/android/usbserial/driver/UsbSerialPort.java
