+++ android/package/src/com/hoho/android/usbserial/driver/UsbSerialProber.java
@@ -1,224 +1,90 @@
-/* Copyright 2011 Google Inc.
+/* Copyright 2011-2013 Google Inc.
+ * Copyright 2013 mike wakerly <opensource@hoho.com>
  *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301,
- * USA.
- *
- * Project home page: http://code.google.com/p/usb-serial-for-android/
+ * Project home page: https://github.com/mik3y/usb-serial-for-android
  */
-
-// IMPORTANT NOTE:
-//  This source has been modified from the original such that testIfSupported only tests for a vendor id
-//  match. If that matches it allows all product ids through. This provides for better match on unknown boards.
 
 package com.hoho.android.usbserial.driver;
 
 import android.hardware.usb.UsbDevice;
-import android.hardware.usb.UsbDeviceConnection;
 import android.hardware.usb.UsbManager;
-import android.util.Log;
 
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
 /**
- * Helper class which finds compatible {@link UsbDevice}s and creates
- * {@link UsbSerialDriver} instances.
- *
- * <p/>
- * You don't need a Prober to use the rest of the library: it is perfectly
- * acceptable to instantiate driver instances manually. The Prober simply
- * provides convenience functions.
- *
- * <p/>
- * For most drivers, the corresponding {@link #probe(UsbManager, UsbDevice)}
- * method will either return an empty list (device unknown / unsupported) or a
- * singleton list. However, multi-port drivers may return multiple instances.
  *
  * @author mike wakerly (opensource@hoho.com)
  */
-public enum UsbSerialProber {
+public class UsbSerialProber {
 
-    // TODO(mikey): Too much boilerplate.
+    private final ProbeTable mProbeTable;
 
-    /**
-     * Prober for {@link FtdiSerialDriver}.
-     *
-     * @see FtdiSerialDriver
-     */
-    FTDI_SERIAL {
-        @Override
-        public List<UsbSerialDriver> probe(final UsbManager manager, final UsbDevice usbDevice) {
-            if (!testIfSupported(usbDevice, FtdiSerialDriver.getSupportedDevices())) {
-                return Collections.emptyList();
-            }
-            final UsbSerialDriver driver = new FtdiSerialDriver(usbDevice);
-            return Collections.singletonList(driver);
-        }
-    },
+    public UsbSerialProber(ProbeTable probeTable) {
+        mProbeTable = probeTable;
+    }
 
-    CDC_ACM_SERIAL {
-        @Override
-        public List<UsbSerialDriver> probe(UsbManager manager, UsbDevice usbDevice) {
-            if (!testIfSupported(usbDevice, CdcAcmSerialDriver.getSupportedDevices())) {
-               return Collections.emptyList();
-            }
-            final UsbSerialDriver driver = new CdcAcmSerialDriver(usbDevice);
-            return Collections.singletonList(driver);
-        }
-    },
-
-    SILAB_SERIAL {
-        @Override
-        public List<UsbSerialDriver> probe(final UsbManager manager, final UsbDevice usbDevice) {
-            if (!testIfSupported(usbDevice, Cp2102SerialDriver.getSupportedDevices())) {
-                return Collections.emptyList();
-            }
-            final UsbSerialDriver driver = new Cp2102SerialDriver(usbDevice);
-            return Collections.singletonList(driver);
-        }
-    },
-
-    PROLIFIC_SERIAL {
-        @Override
-        public List<UsbSerialDriver> probe(final UsbManager manager, final UsbDevice usbDevice) {
-            if (!testIfSupported(usbDevice, ProlificSerialDriver.getSupportedDevices())) {
-                return Collections.emptyList();
-            }
-            final UsbSerialDriver driver = new ProlificSerialDriver(usbDevice);
-            return Collections.singletonList(driver);
-        }
-    };
+    public static UsbSerialProber getDefaultProber() {
+        return new UsbSerialProber(getDefaultProbeTable());
+    }
+    
+    public static ProbeTable getDefaultProbeTable() {
+        final ProbeTable probeTable = new ProbeTable();
+        probeTable.addDriver(CdcAcmSerialDriver.class);
+        probeTable.addDriver(Cp21xxSerialDriver.class);
+        probeTable.addDriver(FtdiSerialDriver.class);
+        probeTable.addDriver(ProlificSerialDriver.class);
+        probeTable.addDriver(Ch34xSerialDriver.class);
+        probeTable.addDriver(GsmModemSerialDriver.class);
+        probeTable.addDriver(ChromeCcdSerialDriver.class);
+        return probeTable;
+    }
 
     /**
-     * Tests the supplied {@link UsbDevice} for compatibility with this enum
-     * member, returning one or more driver instances if compatible.
+     * Finds and builds all possible {@link UsbSerialDriver UsbSerialDrivers}
+     * from the currently-attached {@link UsbDevice} hierarchy. This method does
+     * not require permission from the Android USB system, since it does not
+     * open any of the devices.
      *
-     * @param manager the {@link UsbManager} to use
-     * @param usbDevice the raw {@link UsbDevice} to use
-     * @return zero or more {@link UsbSerialDriver}, depending on compatibility
-     *         (never {@code null}).
+     * @param usbManager usb manager
+     * @return a list, possibly empty, of all compatible drivers
      */
-    protected abstract List<UsbSerialDriver> probe(final UsbManager manager, final UsbDevice usbDevice);
+    public List<UsbSerialDriver> findAllDrivers(final UsbManager usbManager) {
+        final List<UsbSerialDriver> result = new ArrayList<>();
 
-    /**
-     * Creates and returns a new {@link UsbSerialDriver} instance for the first
-     * compatible {@link UsbDevice} found on the bus.  If none are found,
-     * returns {@code null}.
-     *
-     * <p/>
-     * The order of devices is undefined, therefore if there are multiple
-     * devices on the bus, the chosen device may not be predictable (clients
-     * should use {@link #findAllDevices(UsbManager)} instead).
-     *
-     * @param usbManager the {@link UsbManager} to use.
-     * @return the first available {@link UsbSerialDriver}, or {@code null} if
-     *         none are available.
-     */
-    public static UsbSerialDriver findFirstDevice(final UsbManager usbManager) {
         for (final UsbDevice usbDevice : usbManager.getDeviceList().values()) {
-            for (final UsbSerialProber prober : values()) {
-                final List<UsbSerialDriver> probedDevices = prober.probe(usbManager, usbDevice);
-                if (!probedDevices.isEmpty()) {
-                    return probedDevices.get(0);
-                }
+            final UsbSerialDriver driver = probeDevice(usbDevice);
+            if (driver != null) {
+                result.add(driver);
             }
+        }
+        return result;
+    }
+    
+    /**
+     * Probes a single device for a compatible driver.
+     * 
+     * @param usbDevice the usb device to probe
+     * @return a new {@link UsbSerialDriver} compatible with this device, or
+     *         {@code null} if none available.
+     */
+    public UsbSerialDriver probeDevice(final UsbDevice usbDevice) {
+        final Class<? extends UsbSerialDriver> driverClass = mProbeTable.findDriver(usbDevice);
+        if (driverClass != null) {
+            final UsbSerialDriver driver;
+            try {
+                final Constructor<? extends UsbSerialDriver> ctor =
+                        driverClass.getConstructor(UsbDevice.class);
+                driver = ctor.newInstance(usbDevice);
+            } catch (NoSuchMethodException | IllegalArgumentException | InstantiationException |
+                     IllegalAccessException | InvocationTargetException e) {
+                throw new RuntimeException(e);
+            }
+            return driver;
         }
         return null;
     }
 
-    /**
-     * Creates a new {@link UsbSerialDriver} instance for all compatible
-     * {@link UsbDevice}s found on the bus. If no compatible devices are found,
-     * the list will be empty.
-     *
-     * @param usbManager
-     * @return
-     */
-    public static List<UsbSerialDriver> findAllDevices(final UsbManager usbManager) {
-        final List<UsbSerialDriver> result = new ArrayList<UsbSerialDriver>();
-        //Log.i("QGC_UsbSerialProber", "Looking for USB devices");
-        // For each UsbDevice, call probe() for each prober.
-        for (final UsbDevice usbDevice : usbManager.getDeviceList().values()) {
-            //Log.i("QGC_UsbSerialProber", "Probing device: " + usbDevice.getDeviceName() + " mid: " + usbDevice.getVendorId() + " pid: " + usbDevice.getDeviceId());
-            result.addAll(probeSingleDevice(usbManager, usbDevice));
-        }
-        return result;
-    }
-
-    /**
-     * Special method for testing a specific device for driver support,
-     * returning any compatible driver(s).
-     *
-     * <p/>
-     * Clients should ordinarily use {@link #findAllDevices(UsbManager)}, which
-     * operates against the entire bus of devices. This method is useful when
-     * testing against only a single target is desired.
-     *
-     * @param usbManager the {@link UsbManager} to use.
-     * @param usbDevice the device to test against.
-     * @return a list containing zero or more {@link UsbSerialDriver} instances.
-     */
-    public static List<UsbSerialDriver> probeSingleDevice(final UsbManager usbManager,
-            UsbDevice usbDevice)
-    {
-        final List<UsbSerialDriver> result = new ArrayList<UsbSerialDriver>();
-        for (final UsbSerialProber prober : values()) {
-            final List<UsbSerialDriver> probedDevices = prober.probe(usbManager, usbDevice);
-            result.addAll(probedDevices);
-        }
-        return result;
-    }
-
-    /**
-     * Deprecated; Use {@link #findFirstDevice(UsbManager)}.
-     *
-     * @param usbManager
-     * @return
-     */
-    @Deprecated
-    public static UsbSerialDriver acquire(final UsbManager usbManager) {
-        return findFirstDevice(usbManager);
-    }
-
-    /**
-     * Deprecated; use {@link #probeSingleDevice(UsbManager, UsbDevice)}.
-     *
-     * @param usbManager
-     * @param usbDevice
-     * @return
-     */
-    @Deprecated
-    public static UsbSerialDriver acquire(final UsbManager usbManager, final UsbDevice usbDevice) {
-        final List<UsbSerialDriver> probedDevices = probeSingleDevice(usbManager, usbDevice);
-        if (!probedDevices.isEmpty()) {
-            return probedDevices.get(0);
-        }
-        return null;
-    }
-
-    /**
-     * Returns {@code true} if the given device is found in the driver's
-     * vendor/product map.
-     *
-     * @param usbDevice the device to test
-     * @param supportedDevices map of vendor IDs to product ID(s)
-     * @return {@code true} if supported
-     */
-    private static boolean testIfSupported(final UsbDevice usbDevice, final Map<Integer, int[]> supportedDevices) {
-        return supportedDevices.containsKey(usbDevice.getVendorId());
-    }
 }
