+++ android/package/src/org/mavlink/qgroundcontrol/QGCUsbId.java
@@ -1,0 +1,65 @@
+/* Copyright 2011-2013 Google Inc.
+ * Copyright 2013 mike wakerly <opensource@hoho.com>
+ *
+ * Project home page: https://github.com/mik3y/usb-serial-for-android
+ */
+
+package org.mavlink.qgroundcontrol;
+
+/**
+ * Registry of USB vendor/product ID constants.
+ *
+ * Culled from various sources; see
+ * <a href="http://www.linux-usb.org/usb.ids">usb.ids</a> for one listing.
+ *
+ * @author mike wakerly (opensource@hoho.com)
+ */
+public final class QGCUsbId {
+
+    public static final int VENDOR_PX4    = 0x26AC;
+    public static final int DEVICE_PX4FMU = 0x11;
+
+    public static final int VENDOR_ATMEL = 0x03EB;
+    public static final int ATMEL_LUFA_CDC_DEMO_APP = 0x2044;
+
+    public static final int VENDOR_ARDUINO = 0x2341;
+    public static final int ARDUINO_UNO = 0x0001;
+    public static final int ARDUINO_MEGA_2560 = 0x0010;
+    public static final int ARDUINO_SERIAL_ADAPTER = 0x003b;
+    public static final int ARDUINO_MEGA_ADK = 0x003f;
+    public static final int ARDUINO_MEGA_2560_R3 = 0x0042;
+    public static final int ARDUINO_UNO_R3 = 0x0043;
+    public static final int ARDUINO_MEGA_ADK_R3 = 0x0044;
+    public static final int ARDUINO_SERIAL_ADAPTER_R3 = 0x0044;
+    public static final int ARDUINO_LEONARDO = 0x8036;
+
+    public static final int VENDOR_VAN_OOIJEN_TECH = 0x16c0;
+    public static final int VAN_OOIJEN_TECH_TEENSYDUINO_SERIAL = 0x0483;
+
+    public static final int VENDOR_LEAFLABS = 0x1eaf;
+    public static final int LEAFLABS_MAPLE = 0x0004;
+
+    public static final int VENDOR_UBLOX = 0x1546;
+    public static final int DEVICE_UBLOX_5 = 0x01a5;
+    public static final int DEVICE_UBLOX_6 = 0x01a6;
+    public static final int DEVICE_UBLOX_7 = 0x01a7;
+    public static final int DEVICE_UBLOX_8 = 0x01a8;
+
+    public static final int VENDOR_OPENPILOT = 0x20A0;
+    public static final int DEVICE_REVOLUTION = 0x415E;
+    public static final int DEVICE_OPLINK = 0x415C;
+    public static final int DEVICE_SPARKY2 = 0x41D0;
+    public static final int DEVICE_CC3D = 0x415D;
+
+    public static final int VENDOR_ARDUPILOT_CHIBIOS1 = 0x0483;
+    public static final int VENDOR_ARDUPILOT_CHIBIOS2 = 0x1209;
+    public static final int DEVICE_ARDUPILOT_CHIBIOS =  0x5740;
+
+    public static final int VENDOR_DRAGONLINK = 0x1FC9;
+    public static final int DEVICE_DRAGONLINK = 0x0083;
+
+    private UsbId() {
+        throw new IllegalAccessError("Non-instantiable class");
+    }
+
+}
